requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
darker==1.5.1
flake8==5.0.4
frozenlist==1.3.3
idna==3.4
importlib-metadata==4.13.0
isort==5.10.1
mccabe==0.7.0
msgpack==1.0.4
multidict==6.0.2
mypy-extensions==0.4.3
nanoid==2.0.0
pathspec==0.10.2
pipdeptree==2.3.3
platformdirs==2.5.4
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.13.0
pyperclip==1.8.2
python-dotenv==0.21.0
requests==2.28.1
rich==12.6.0
textual==0.4.0
toml==0.10.2
tomli==2.0.1
typing-extensions==4.4.0
urllib3==1.26.12
yarl==1.8.1
zipp==3.10.0
